修改了翻译漏翻逻辑,单条翻译采用线程池,token动态计算,不在手动分割.

+++ src/main/java/com/xhxj/jsongpttranslator/service/chatgpt/ChatGptTranslationAsyncService.java
@@ -18,5 +18,10 @@ public interface ChatGptTranslationAsyncService {
      */
     CompletableFuture<Void> accessingChatGptOne(TranslationData translationData);
 
-    Integer calculateToken(List<TranslationData> translationData);
+    /**
+     * 获取token
+     * @param translationData
+     * @return
+     */
+    Long calculateToken(List<TranslationData> translationData);
 }
